Changed column type from 'bool' to 'boolean'.
This was most probably just a spelling error from @ad9b1948, but it made
Rails unable to create a proper schema file, leading to me being unable
to prepare a test database from it.

This is a correction to an already existing migration, which means that
all developers would have to reset their database by running 'rake
db:migrate:reset' and presumably reseeding it with 'rake db:seed'.