Added Annotate and annotations to all models.
Annotate is a gem which injects each model with a comment block
containing the actual columns representing the model in the database.
This sort of serves as an inline documentation and is somewhat useful.

The annotate gem will add an executable called 'annotate' in your path,
asumming you are using RVM or something equivalent, and running this
will update each model any changes made since last time.