Removing the accepcted_privacy column from User.
A model in Rails can have an attribute that does not exists in the
database by using #attr_accesssor, which will dynamically create
accessor methods for an instance variable with the given name.

Since a user must accept the privacy terms upon registrations, the only
value one is ever going to see in the database is 'true', making it
obviously not necessary to store. Everything behaves like it did before,
except that the true-value is not stored in the database.

See http://apidock.com/ruby/Module/attr_accessor for more information.