Rewriting all actions of info to Markdown.
Markdown parsing is currently done using the gem rdiscount.

Discount is an implementation of John Gruber's Markdown markup language
in C. It implements all of the language described in the markdown syntax
document and passes the Markdown 1.0 test suite.

The purpose of Markdown is to make ducuments easy-to-read and
easy-to-write. It is far better than Haml for inline content markup. See
http://goo.gl/kIh3Y for someone who shares my opinion.

I also took the liberty to introduce a 120 character line width
constraint on textual inline content such as what is seen in these
pages.