Cleaning up relation between Order and User.
An 'owned_order' association has been added to User, which works on the
owner_user_id column of Order. The creation of a default order has been
moved from the controller action orders#show into the User model, where
it belongs. This is cleaner and it simplifies the controller code.