Updated speaker registration and added some information.
When a "not logged in" user wants to submit a talk we should
direct him to the "sign up page" instead of the "log in page".
The reason for this is that we need to add a "magic" parameter
to the sign up page to allow the user to select the
"speaker ticket".

I have also added an email confirmation sent to the user and
an notification email sent to us, which fires when a new
speaker signs up.

(I will add cucumber for this later when i have time)