This implements the basic fundations for managing the program, and is based on the visual model used for Smidig 2012 [1], with a more suited fundamental model.
The program basicly has many timeslots. A Timeslot has
a *start* and *end* datetime and consits one or more
roomslots. A roomslot the the actual visual component,
tied to a time-period a room and conists of one or more
talks.

This commit only intoduces all the resources and their
associations. Later commits will add the required
functionality to:

1. build the program (command)
2. present the program (query)

[1] http://participate.smidig2012.no/program